Make visit decltype(auto) for genericity
It should defer the return type (including value category) to the
visitor. It won't matter in most cases, but you might have to return a
reference calculated from any of the ptrs that could be in the variant.
That should be allowed.

Because value categories are subtle, I made an impl::garbage_value
function that will give you a value of the type + value category of
exactly what it was templated on. visit_nth delegates its (fake) base
case to it. For now it calls the default constructor. I think the fully
"correct" behavior would be to dereference nullptr. It will always
typecheck. And this code will only ever be executed if visit_nth is
called horribly out of contract; there is no sensible return value. That
code will be unreachable with assertions enabled anyway.