Add static by value versions of member fns
This way it doesn't compile to a function that accepts a this pointer to a
uintptr_pair/tagged_ptr/variant_ptr. This means the optimizer does not
need to reason about memory when inlining etc these functions.

This is probably a premature optimization but I don't think it makes the
code more complicated.