Refactor cmake to better follow pitchfork layout
It is important to me to separate project definition from the specific
build instructions. How is this to be expressed? Options? Extra
includes? A wrapper project?

For a library, this is especially important.

Before I used Mateusz Pusz's layout (from his CppCon talk "Git, Cmaake,
Conan...") where the toplevel directory had build settings and the src/
directory contained the core definition of the project. This feels
inside-out to me. Now the core definition is in CMakeLists.txt.

Solution for now:
An optional early_hooks.cmake file will be run if found. The one
supplied in this repo is for CI+development. More work needs to be done

This verison of early_hooks loads conan's paths for find_package and
sets high warning levels and fuzzing.